use event dispatcher to delete an alert in parent component
Thanks @RafyAmgadBenjamin :)
